Add multi-stage Dockerfile to build alpine image
alpine.Dockerfile builds all the dependencies with python:3.6 as a base image and then copies
the binaries into an python3.6 alpine image. The resulting image size is now 230MB instead of over 1GB.

The make.sh script is not needed anymore by the new Dockerfile.
In addition the autobuild feature of Dockerhub also builds and pushes this image to teambitflow/anomaly-injector-agent:master
on every new commit.